Fix compound face tesselation.
