+++ Elements/src/Geometry/Solids/Solid.cs
@@ -9,6 +9,7 @@ using System.Text;
 using Elements.Geometry.Interfaces;
 using Elements.Search;
 using Elements.Spatial;
+using LibTessDotNet.Double;
 
 [assembly: InternalsVisibleTo("Hypar.Elements.Tests")]
 
@@ -847,6 +848,75 @@ namespace Elements.Geometry.Solids
             {
                 e.Right.Loop.InsertEdgeBefore(e.Right, e1.Right);
             }
+        }
+
+        internal Csg.Solid ToCsg()
+        {
+            var polygons = new List<Csg.Polygon>();
+
+            foreach (var f in Faces.Values)
+            {
+                var tess = new Tess
+                {
+                    NoEmptyPolygons = true
+                };
+
+                tess.AddContour(f.Outer.ToContourVertexArray());
+
+                if (f.Inner != null)
+                {
+                    foreach (var loop in f.Inner)
+                    {
+                        tess.AddContour(loop.ToContourVertexArray());
+                    }
+                }
+
+                tess.Tessellate(WindingRule.Positive, ElementType.Polygons, 3);
+
+                Vector3 e1 = new Vector3();
+                Vector3 e2 = new Vector3();
+
+                var vertices = new List<Csg.Vertex>(tess.Elements.Count() * 3);
+                for (var i = 0; i < tess.ElementCount; i++)
+                {
+                    var a = tess.Vertices[tess.Elements[i * 3]].ToCsgVector3();
+                    var b = tess.Vertices[tess.Elements[i * 3 + 1]].ToCsgVector3();
+                    var c = tess.Vertices[tess.Elements[i * 3 + 2]].ToCsgVector3();
+
+                    Csg.Vertex av = null;
+                    Csg.Vertex bv = null;
+                    Csg.Vertex cv = null;
+
+                    if (i == 0)
+                    {
+                        var n = f.Plane().Normal;
+                        e1 = n.Cross(n.IsParallelTo(Vector3.XAxis) ? Vector3.YAxis : Vector3.XAxis).Unitized();
+                        e2 = n.Cross(e1).Unitized();
+                    }
+                    if (av == null)
+                    {
+                        var avv = new Vector3(a.X, a.Y, a.Z);
+                        av = new Csg.Vertex(a, new Csg.Vector2D(e1.Dot(avv), e2.Dot(avv)));
+                        vertices.Add(av);
+                    }
+                    if (bv == null)
+                    {
+                        var bvv = new Vector3(b.X, b.Y, b.Z);
+                        bv = new Csg.Vertex(b, new Csg.Vector2D(e1.Dot(bvv), e2.Dot(bvv)));
+                        vertices.Add(bv);
+                    }
+                    if (cv == null)
+                    {
+                        var cvv = new Vector3(c.X, c.Y, c.Z);
+                        cv = new Csg.Vertex(c, new Csg.Vector2D(e1.Dot(cvv), e2.Dot(cvv)));
+                        vertices.Add(cv);
+                    }
+
+                    var p = new Csg.Polygon(new List<Csg.Vertex>() { av, bv, cv });
+                    polygons.Add(p);
+                }
+            }
+            return Csg.Solid.FromPolygons(polygons);
         }
     }
 }
